+++ LabXml/Azure/AzureRmVmSize.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 
 namespace AutomatedLab.Azure
 {
@@ -11,9 +12,11 @@ namespace AutomatedLab.Azure
         public int? MaxDataDiskCount { get; set; }
         public int ResourceDiskSizeInMB { get; set; }
         public int OSDiskSizeInMB { get; set; }
+        public bool Gen1Supported {get; set;}
+        public bool Gen2Supported {get; set;}
 
         public AzureRmVmSize()
-        { }
+        {  }
 
         public override string ToString()
         {